pyproject.toml
[build-system]
requires = ["setuptools","pyyaml"]
build-backend = "setuptools.build_meta"

[project]
name = "pyNetTestShell"
description = "An interactive python shell to test the network."
readme = "README.md"
requires-python = ">=3.11"
version = "2.2.0-alpha"
classifiers = [
    "Development Status::5 Production/Stable",
    "Environment::Console",
    "Intended Audience::System Administrators",
    "Intended Audeience::Telecommunications Industry",
    "License::OSI Approved::The Unlicense (Unlicense)",
    "Natural Language::English",
    "Operating System::POSIX::Linux",
    "Topic::System::Network::Monitoring"
]
dependencies = [
    "dnspython",
    "PyYAML",
    "pandas",
    "speedtest-cli",
    "numpy",
    "fire",
    "tqdm"
]

[project.urls]
"Homepage" = "https://github.com/packetpunter/pyNetTestShell"
